Deduplicate scoring test catalog

Copilot-Session: f42917e2-d298-4bde-9056-79a9e6e17dfa
